models/rakuten-llm/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
llama-cpp-python>=0.2.0
numpy>=1.24.0
python-dotenv>=1.0.0